+++ lib/isolates/isolate_factory.dart
@@ -9,6 +9,7 @@ import 'isolate_bme280.dart';
 import 'isolate_bme680.dart';
 import 'isolate_cozir.dart';
 import 'isolate_demo.dart';
+import 'isolate_hat_adc.dart';
 import 'isolate_helper.dart';
 import 'isolate_leds.dart';
 import 'isolate_mcp9808.dart';
@@ -17,6 +18,7 @@ import 'isolate_sdc30.dart';
 import 'isolate_sgp30.dart';
 import 'isolate_sht31.dart';
 import 'isolate_si1145.dart';
+import 'isolate_tsl2591.dart';
 
 /// Constructs a class by name.
 class IsolateClassFactory {
@@ -56,6 +58,12 @@ class IsolateClassFactory {
     SI1145isolate.empty().runtimeType.toString():
         (String isolateId, Object data) =>
             SI1145isolate(isolateId, data as bool),
+    HatADCisolate.empty().runtimeType.toString():
+        (String isolateId, Object data) =>
+            HatADCisolate(isolateId, data as bool),
+    TSL2591isolate.empty().runtimeType.toString():
+        (String isolateId, Object data) =>
+            TSL2591isolate(isolateId, data as bool),
   };
 
   static IsolateWrapper createInstance(