+++ lib/main.dart
@@ -10,9 +10,11 @@ import 'package:isolate/components/error_box.dart';
 import 'package:isolate/dashboards/dashboard_bme680.dart';
 import 'package:isolate/dashboards/dashboard_cozir.dart';
 import 'package:isolate/dashboards/dashboard_gesture.dart';
+import 'package:isolate/dashboards/dashboard_hat_adc.dart';
 import 'package:isolate/dashboards/dashboard_leds.dart';
 import 'package:isolate/dashboards/dashboard_overview.dart';
 import 'package:isolate/isolates/isolate_cozir.dart';
+import 'package:isolate/isolates/isolate_hat_adc.dart';
 import 'package:isolate/isolates/isolate_leds.dart';
 import 'package:isolate/tab/about_tab.dart';
 import 'package:isolate/tab/configuration_tab.dart';
@@ -29,6 +31,7 @@ import 'dashboards/dashboard_sdc30.dart';
 import 'dashboards/dashboard_sgp30.dart';
 import 'dashboards/dashboard_sht31.dart';
 import 'dashboards/dashboard_si1145.dart';
+import 'dashboards/dashboard_tsl2591.dart';
 import 'isolates/isolate_bme280.dart';
 import 'isolates/isolate_bme680.dart';
 import 'isolates/isolate_demo.dart';
@@ -40,9 +43,10 @@ import 'isolates/isolate_sdc30.dart';
 import 'isolates/isolate_sgp30.dart';
 import 'isolates/isolate_sht31.dart';
 import 'isolates/isolate_si1145.dart';
+import 'isolates/isolate_tsl2591.dart';
 
 void main() async {
-  // Set the size of the desktop app to a fixed size to enable
+  // Sets the size of the desktop app to a fixed size to enable making
   // screenshots with the same size.
   if (Platform.isWindows || Platform.isMacOS) {
     WidgetsFlutterBinding.ensureInitialized();
@@ -161,6 +165,13 @@ Widget _createDashboard(ParentUpdateCallback callback) {
       );
     case DashboardType.overview:
       return DashboardOverview(callback: callback);
+    case DashboardType.adc:
+      return DashboardHatADC(
+          isolateWrapper:
+              HatADCisolate(DashboardType.adc.name, gSimulateSensor));
+    case DashboardType.tsl2591:
+      return DashboardTSL2591(
+          isolateWrapper: TSL2591isolate(gDashboard.name, gSimulateSensor));
   }
 }