+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## 1.0.2
+## 1.0.3
+* Add ADC (Analog Digital Converter) demo
+* Add [TSL2591](https://www.adafruit.com/product/1980) sensor
 
-* Add [SI1145]( https://www.seeedstudio.com/Grove-Sunlight-Sensor.html) sunlight sensor: visible, IR and raw UV
+## 1.0.2
+* Add [SI1145](https://www.seeedstudio.com/Grove-Sunlight-Sensor.html) sunlight sensor: visible, IR and raw UV
 
 ## 1.0.1 
-
 * Add [SDC30](https://www.seeedstudio.com/Grove-CO2-Temperature-Humidity-Sensor-SCD30-p-2911.html) - CO2, temperature amd humidity sensor